# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Farewell 2026 Event Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /voting/status:
    get:
      operationId: getVotingStatus
      tags: [voting]
      summary: Get voting status
      responses:
        "200":
          description: Voting status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VotingStatus"
    put:
      operationId: setVotingStatus
      tags: [voting]
      summary: Enable/disable voting (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetVotingStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VotingStatus"
  /voting/audience:
    post:
      operationId: submitAudienceVote
      tags: [voting]
      summary: Submit audience vote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudienceVoteRequest"
      responses:
        "200":
          description: Vote recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Already voted or voting disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /voting/faculty:
    post:
      operationId: submitFacultyScore
      tags: [voting]
      summary: Submit faculty score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FacultyScoreRequest"
      responses:
        "200":
          description: Score recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Already scored or voting disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /rampwalk/next:
    post:
      operationId: nextParticipant
      tags: [rampwalk]
      summary: Move to next participant (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NextParticipantRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RampWalkState"
  /rampwalk/settings:
    put:
      operationId: setRampwalkSettings
      tags: [rampwalk]
      summary: Update rampwalk settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RampWalkSettingsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RampWalkState"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Participant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        categoryId:
          type: integer
        categoryName:
          type: string
        contestantNo:
          type: integer
        photoUrl:
          type: string
          nullable: true
        orderIndex:
          type: integer
      required: [id, name, gender, categoryId, contestantNo, orderIndex]
...
    VotingStatus:
      type: object
      properties:
        isOpen:
          type: boolean
        activeCategoryId:
          type: integer
          nullable: true
        activeCategoryName:
          type: string
          nullable: true
      required: [isOpen]
    SetVotingStatusRequest:
      type: object
      properties:
        isOpen:
          type: boolean
        activeCategoryId:
          type: integer
          nullable: true
      required: [isOpen]
    AudienceVoteRequest:
      type: object
      properties:
        participantId:
          type: integer
        categoryId:
          type: integer
      required: [participantId, categoryId]
    FacultyScoreRequest:
      type: object
      properties:
        participantId:
          type: integer
        scoreIntroduction:
          type: integer
          minimum: 1
          maximum: 5
        scoreRampwalk:
          type: integer
          minimum: 1
          maximum: 5
        scoreTalent:
          type: integer
          minimum: 1
          maximum: 5
      required: [participantId]
    RampWalkState:
      type: object
      properties:
        currentParticipantId:
          type: integer
          nullable: true
        isLive:
          type: boolean
        ratingActive:
          type: boolean
        participant:
          $ref: "#/components/schemas/Participant"
          nullable: true
      required: [currentParticipantId, isLive, ratingActive]
    RampWalkSettingsRequest:
      type: object
      properties:
        isLive:
          type: boolean
        ratingActive:
          type: boolean
    NextParticipantRequest:
      type: object
      properties:
        participantId:
          type: integer
          nullable: true
      required: [participantId]
...